agsac_dog_navigation/requirements.txt
# 核心深度学习框架
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 数值计算和数据处理
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# 配置管理
pyyaml>=6.0
omegaconf>=2.3.0

# 可视化和日志
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
wandb>=0.15.0

# 几何计算
shapely>=2.0.0
scikit-learn>=1.3.0

# 强化学习环境
gymnasium>=0.28.0
stable-baselines3>=2.0.0

# 进度条和工具
tqdm>=4.65.0
rich>=13.0.0

# 测试框架
pytest>=7.3.0
pytest-cov>=4.1.0

# 代码质量
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Jupyter支持
jupyter>=1.0.0
ipykernel>=6.25.0

# 可选：Gazebo仿真支持
# gazebo-ros>=3.0.0  # 需要ROS环境
# rospy>=1.15.0      # 需要ROS环境
